Improve invoice details layout and fix Days Outstanding display - Convert Days Outstanding to integer (no decimals) - Redesign Invoice Details section with better layout and spacing - Add Payment Progress card with visual progress bar - Improve Financial Summary layout with better visual hierarchy - Prevent content overflow with proper spacing and borders

+++ app/Http/Controllers/AccountsReceivableController.php
@@ -1668,7 +1668,8 @@ class AccountsReceivableController extends Controller
             // Calculate statistics
             $totalPayments = $invoicePayments->sum('amount');
             $paymentCount = $invoicePayments->count();
-            $daysOutstanding = $invoice->due_date ? now()->diffInDays($invoice->due_date, false) : 0;
+            // Calculate days outstanding as integer (no decimals)
+            $daysOutstanding = $invoice->due_date ? (int)now()->diffInDays($invoice->due_date, false) : 0;
 
             // Get activity logs if available
             $activityLogs = [];
